Move mercadopago checkout tracking to extension
